backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
pymysql==1.1.1
cryptography==43.0.0
finmind==0.4.12
feedparser==6.0.11
pandas==2.2.3
numpy==1.26.4
openai>=1.30.0
apscheduler==3.10.4
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
python-multipart==0.0.9
httpx==0.27.0
python-dateutil>=2.8.2
requests>=2.31.0
